+++ browser/interceptor.js
@@ -1,17 +1,45 @@
 'use strict';
 
+var emitter = require('contra.emitter');
 var router = require('./router');
-var interceptors = {};
+var interceptors = emitter();
+
+function getInterceptorEvent (url, route) {
+  var e = {
+    url: url,
+    route: route,
+    parts: route.parts,
+    model: null,
+    defaultPrevented: false,
+    preventDefault: preventDefault
+  };
+
+  function preventDefault (model) {
+    e.defaultPrevented = true;
+    e.model = model;
+  }
+
+  return e;
+}
+
+function add (action, fn) {
+  if (arguments.length === 1) {
+    fn = action;
+    action = '*';
+  }
+  interceptors.on(action, fn);
+}
+
+function execute (route) {
+  var e = getInterceptorEvent(route.url, route);
+
+  interceptors.emit('*', e);
+  interceptors.emit(route.action, e);
+
+  return e;
+}
 
 module.exports = {
-  on: function (action, fn) {
-    interceptors[action] = fn;
-  },
-  intercept: function (url) {
-    var route = router(url);
-    var interceptor = interceptors[route.action];
-    if (interceptor) {
-      return interceptor(route.params);
-    }
-  }
+  add: add,
+  execute: execute
 };